Test mockable functions of UartBoot to increase coverage
This file exists to provide coverage for functions in
uartboot_testing_implementations.cpp

UartBoot uses function primities for interacting with
the EEPROM and Flash.

These primitives are part of
https://github.com/mihaigalos/avr-bootloader-common,
however, in order to test the logic of UartBoot,
they are mocked by using virtual methods with the
same name in class UartBoot.